Rewrite targets file parser
I thought the file format was needlessly fussy about what it accepted
and rewrote the parser to be more lenient.

The new parser is line-based instead of byte-based (which seems
reasonable, as the target file format is line based), and accepts the
following syntax:

    HTTP_METHOD url
    $ body

The body line is optional. The rules for what is considered to be a body
line are:

    (1) If something starts with '$ ', it's a body
    (2) If the line is literally '{}', it's an empty body

Point (2) is there to be backwards-compatible with the original format.

A missing body line is taken to mean an empty request body. For example,
this is now fine and signifies requests with an empty body:

    GET http://www.example.com
    POST localhost:4000
    DELETE http://127.0.0.1

We also support non-GET HTTP methods now as a consequence of the parser
rewrite.