add bool command line parameter -randomOrder
randomOrder tries to avoid cache hits when executing
multiple instances of slapper in parallel.

with sufficient lines in the targets file, you could guarantee
none of the requests would hit a cache, but if you run multiple
instances of slapper in parallel one instance could hit the cache
entry created by the other instance.

with randomOrder, the two instances are less likely to create
cache entries for each other.